refactoring of XmlService
