+++ MysteryBox.Web/Services/Common/IXmlService.cs
@@ -1,6 +1,6 @@
-﻿namespace MysteryBox.WebService.Services
+﻿namespace MysteryBox.WebService.Services.Common
 {
-    public interface IDomainboxXmlService
+    public interface IXmlService
     {
         string Serialize<T>(T payload) where T : class;
         T DeserializeXml<T>(string content) where T : class;
